requirements.txt
fastapi
uvicorn
celery
redis
b2sdk
python-multipart
pydantic
aiofiles
httpx
loguru
flower
python-dotenv